Move machines inside workflow module
